int/builder/markdown: improves code organization with sub packages
