+++ internal/builder/markdown/hook/hookie.go
@@ -1,0 +1,29 @@
+package hook
+
+import (
+	"io"
+
+	"github.com/gomarkdown/markdown/ast"
+	"github.com/ufukty/kask/internal/builder/markdown/hook/codefence"
+)
+
+type visitor struct {
+	cf *codefence.Renderer
+}
+
+func NewVisitor() *visitor {
+	return &visitor{}
+}
+
+func (v visitor) Visit(w io.Writer, node ast.Node, entering bool) (ast.WalkStatus, bool) {
+	switch node := node.(type) {
+	case *ast.CodeBlock:
+		return v.cf.RenderNodeHook(w, node, entering)
+	case *ast.Image:
+		// TODO: change destination
+
+	case *ast.Link:
+		links(node)
+	}
+	return ast.GoToNext, false
+}
